Link to GoDaddy help page for Glue Record setup (called Host Names)
I struggled for half a day to find any mention of how to set up Glue
Records on GoDaddy -- mainly because there's no mention of the term on
their website.

Turns out that they refer to these as "Host Names" (a rather generic
term that doesn't help when you search for it either).

Adding a link to a page that is at somewhat helpful in locating this.