Deleting the contents of the /home/user-data/ssl (#95)
* Update maintenance.html

When a new box is created the ssl data directory contains a self-signed certificate.  When the backup is restored, this data then conflicts with the backed up data causing nginx to not restart.  Deleting the contents of the /home/user-data/ssl directory will solve this issue.

* Update guide.html (fix typo)

Fix typo in setup guide referring to Ubuntu 18.04 rather than 22.04.