Update maintenance.html
The instructions for restoring from S3 compatible backups did not work, so I updated and enhanced the instructions. They now work, and are clearer and more complete. Since I was editing the maintenance.html file, I decided to update it more comprehensively. This included updating the body to use UTF-8 character encoding, as specified in the header.