Update maintenance.html
The instructions in the Mail-in-a-Box Maintenance Guide for restoring from S3 compatible backups did not work, so I updated and enhanced the instructions in a forked copy of the maintenance.html file. The instructions now work, and are clearer and more complete. Since I was editing the maintenance.html file, I decided to update it more comprehensively. This included updating the body to use UTF-8 character encoding, as specified in the header. I also updated the instructions for maintaining Ubuntu Linux to be easier for newer users who are still learning and becoming more comfortable with system maintenance.